Nanhu 平台Verl适配
DeepSeek-V2-Lite, actor & ref 使用EP8, rollout 使用TP8EP8跑通

备份对应修改和测试脚本

+++ verl/models/mcore/config_converter.py
@@ -423,7 +423,8 @@ def hf_to_mcore_config_dpskv2(
         moe_router_topk_scaling_factor=hf_config.routed_scaling_factor,
         moe_layer_freq=moe_layer_freq,
         # mcore 0.12 moe
-        moe_router_dtype="fp64",
+        # moe_router_dtype="fp64",
+        moe_router_dtype="fp32", # ATTN 修改 router 类型
         disable_bf16_reduced_precision_matmul=True,
         # Other optimizations
         # deallocate_pipeline_outputs=True,
